initial architechture setup and clean up
