+++ app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/MainActivity.kt
@@ -3,7 +3,6 @@ package io.github.viabachelora23michaelkutaibakasper.bprapp
 import android.Manifest
 import android.content.pm.PackageManager
 import android.os.Bundle
-import android.util.Log
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.result.contract.ActivityResultContracts
@@ -11,8 +10,6 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material.icons.filled.Build
@@ -21,7 +18,6 @@ import androidx.compose.material.icons.filled.Edit
 import androidx.compose.material3.BottomAppBar
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
@@ -30,10 +26,8 @@ import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
@@ -41,17 +35,7 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.core.content.ContextCompat
-import com.google.android.gms.maps.model.CameraPosition
-import com.google.android.gms.maps.model.LatLng
-import com.google.maps.android.compose.GoogleMap
-import com.google.maps.android.compose.MapProperties
-import com.google.maps.android.compose.MapUiSettings
-import com.google.maps.android.compose.Marker
-import com.google.maps.android.compose.MarkerState
-import com.google.maps.android.compose.rememberCameraPositionState
-import io.github.viabachelora23michaelkutaibakasper.bprapp.data.ApolloCountryClient
-import io.github.viabachelora23michaelkutaibakasper.bprapp.domain.CountryClient
-import io.github.viabachelora23michaelkutaibakasper.bprapp.domain.SimpleCountry
+import io.github.viabachelora23michaelkutaibakasper.bprapp.ui.screens.events.MapListView
 import io.github.viabachelora23michaelkutaibakasper.bprapp.ui.theme.BPRAppTheme
 
 class MainActivity : ComponentActivity() {
@@ -92,7 +76,6 @@ class MainActivity : ComponentActivity() {
                     modifier = Modifier.fillMaxSize(),
                     color = MaterialTheme.colorScheme.background
                 ) {
-                    loadEventList()
 
                     MainScreen()
                 }
@@ -160,109 +143,19 @@ fun MainScreen(modifier: Modifier = Modifier) {
             verticalArrangement = Arrangement.spacedBy(16.dp),
         ) {
             // Map()
-            CountryList()
+            val MapListView = MapListView()
+            MapListView.EventList()
+          //  CountryList()
             // CountrySpecific(code ="DK")
         }
     }
 }
 
-@Composable
-fun Map() {
-    val horsens = LatLng(55.862207, 9.844651)
-    val cameraPositionState = rememberCameraPositionState {
-        position = CameraPosition.fromLatLngZoom(horsens, 15f)
-    }
-    val uiSettings by remember {
-        mutableStateOf(
-            MapUiSettings(
-                myLocationButtonEnabled = true,
-                zoomControlsEnabled = false,
-                compassEnabled = true,
-                mapToolbarEnabled = true,
-                rotationGesturesEnabled = true, tiltGesturesEnabled = true
-            )
-        )
-    }
-    GoogleMap(
-        modifier = Modifier.fillMaxSize(),
-        cameraPositionState = cameraPositionState,
-        uiSettings = uiSettings,
-        properties = MapProperties(isMyLocationEnabled = true)
-    ) {
-        eventList.forEach {
-            Marker(
-                state = MarkerState(position = it.location),
-                title = it.name,
-                snippet = "description"
 
-            )
-        }
 
-    }
-}
 
-fun loadEventList() {
-    eventList.add(event1)
-    eventList.add(event2)
-    eventList.add(event3)
-}
 
-class Event(val name: String, val location: LatLng)
 
-val eventList = mutableListOf<Event>()
-var event1 = Event("Run event", LatLng(55.862207, 9.844651))
-var event2 = Event("Dance event", LatLng(55.872207, 9.744651))
-var event3 = Event("Drunk event", LatLng(55.882207, 9.644651))
-
-@Composable
-fun CountryList() {
-    var response by remember { mutableStateOf<List<SimpleCountry>>(emptyList()) }
-    LaunchedEffect(Unit) {
-        val countryClient: CountryClient = ApolloCountryClient()
-        try {
-            val countries = countryClient.getCountries("https://countries.trevorblades.com/")
-            response = countries
-
-        } catch (e: Exception) {
-            Log.d("countriesList", "Failure", e)
-        }
-    }
-    LazyColumn {
-        items(response) { country ->
-            Column {
-                Text("Country: ${country.name ?: "No name"}")
-                Text(text = "Country code: ${country.code ?: "No code"}")
-                Text(text = "Capital: ${country.capital ?: "No capital"}")
-                HorizontalDivider()
-            }
-
-        }
-    }
-}
-
-@Composable
-fun CountrySpecific(code: String) {
-    var response by remember { mutableStateOf<List<SimpleCountry>>(mutableListOf()) }
-    LaunchedEffect(Unit) {
-        val countryClient: CountryClient = ApolloCountryClient()
-        try {
-            val countries = countryClient.getCountries(url="https://countries.trevorblades.com/")
-            if (countries != null) {
-                response = countries
-            }
-
-        } catch (e: Exception) {
-            Log.d("countriesList", "Failure", e)
-        }
-    }
-    Column {
-
-        Text(text = "Country: ${response[0].name}")
-        Text(text = "Country code: ${response[0].code ?: "No code"}")
-        Text(text = "Capital: ${response[0].capital ?: "No capital"}")
-        HorizontalDivider()
-    }
-}
 
 @Preview(showBackground = true)
 @Composable
