+++ app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/data/domain/Event.kt
@@ -1,0 +1,7 @@
+package io.github.viabachelora23michaelkutaibakasper.bprapp.data.domain
+
+data class Event(
+    val name:String,
+    val capital: String?,
+    val code: String?
+)
