+++ app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/data/repository/EventRepository.kt
@@ -1,40 +1,43 @@
-package io.github.viabachelora23michaelkutaibakasper.bprapp.data
+package io.github.viabachelora23michaelkutaibakasper.bprapp.data.repository
 
 import android.util.Log
 import com.apollographql.apollo3.ApolloClient
 import io.github.viabachelora23michaelkutaibakasper.bprapp.CountryByCodeQuery
 import io.github.viabachelora23michaelkutaibakasper.bprapp.ExampleQuery
-import io.github.viabachelora23michaelkutaibakasper.bprapp.domain.CountryClient
-import io.github.viabachelora23michaelkutaibakasper.bprapp.domain.SimpleCountry
+import io.github.viabachelora23michaelkutaibakasper.bprapp.data.domain.Event
 
-class ApolloCountryClient() : CountryClient {
-    override suspend fun getCountries(url:String): List<SimpleCountry> {
-       val apolloClient = ApolloClient.Builder()
-            .serverUrl(url)
+class EventRepository : IEventRepository {
+
+    private val URL = "https://countries.trevorblades.com/"
+    override suspend fun getEvents(): List<Event> {
+        val apolloClient = ApolloClient.Builder()
+            .serverUrl(URL)
             .build()
         val response = apolloClient.query(ExampleQuery()).execute()
         Log.d("ApolloCountryClient", "getCountries: ${response.data?.countries}")
         return response.data?.countries?.map {
-            SimpleCountry(
+            Event(
                 it.name,
                 it.capital,
-                it.code)
+                it.code
+            )
         } ?: emptyList()
     }
 
-    override suspend fun getCountry(url: String, code: String): SimpleCountry {
-       val apolloClient = ApolloClient.Builder()
-            .serverUrl(url)
+    override suspend fun getEvent(code: String): Event? {
+        val apolloClient = ApolloClient.Builder()
+            .serverUrl(URL)
             .build()
         val response = apolloClient.query(CountryByCodeQuery(code)).execute()
         Log.d("ApolloCountryClient", "getCountry: $response")
         return response.data?.country?.let {
-            SimpleCountry(
+            Event(
                 it.name,
                 it.capital,
                 it.code
             )
         } ?: throw Exception("Country not found with country code: $code")
     }
+
 
 }
