+++ app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/data/repository/IEventRepository.kt
@@ -1,0 +1,8 @@
+package io.github.viabachelora23michaelkutaibakasper.bprapp.data.repository
+
+import io.github.viabachelora23michaelkutaibakasper.bprapp.data.domain.Event
+
+interface IEventRepository {
+    suspend fun getEvents(): List<Event>
+    suspend fun getEvent(code: String): Event?
+}
