--- app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/domain/SimpleCountry.kt
@@ -1,7 +1,0 @@
-package io.github.viabachelora23michaelkutaibakasper.bprapp.domain
-
-data class SimpleCountry(
-    val name:String,
-    val capital: String?,
-    val code: String?
-)
