--- app/src/main/java/io/github/viabachelora23michaelkutaibakasper/bprapp/presentation/CountriesViewModel.kt
@@ -1,6 +1,0 @@
-package io.github.viabachelora23michaelkutaibakasper.bprapp.presentation
-
-import androidx.lifecycle.ViewModel
-
-class CountriesViewModel:ViewModel() {
-}
